requirements.txt
pandas>=2.0
numpy>=1.25
scikit-learn>=1.5
requests>=2.31
PyYAML>=6.0
playwright>=1.47
python-dotenv>=1.0
beautifulsoup4>=4.12
